Fleshed out the user interface
Added support for -h, instead of just --help

change the --count option to --wordcount, then added a backup user
prompt for when the user doesn't use the use this option. A case
could be made for using a positional argument instead, but I'm
pretty happy with this UX flow as it allows for quick execution
for experienced users as well as discoverability for novices.

Added a program description to the help prompt